backend/requirements.txt
asgiref==3.8.1
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-utils2==2.12.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
# psycopg2==2.9.5
psycopg2-binary
PyJWT==2.6.0
python-utils==3.8.2
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
